+++ AkariRenderer/Src/RHI/IndexBuffer.cpp
@@ -1,6 +1,6 @@
 #include "pch.h"
 
-#include "IndexBuffer.h">"
+#include "IndexBuffer.h"
 
 using namespace Akari;
 
@@ -27,7 +27,7 @@ IndexBuffer::IndexBuffer( Device& device, Microsoft::WRL::ComPtr<ID3D12Resource>
 
 void IndexBuffer::CreateIndexBufferView()
 {
-    UINT bufferSize = m_NumIndices * ( m_IndexFormat == DXGI_FORMAT_R16_UINT ? 2 : 4 );
+    UINT bufferSize = static_cast<UINT>(m_NumIndices) * ( m_IndexFormat == DXGI_FORMAT_R16_UINT ? 2 : 4 );
 
     m_IndexBufferView.BufferLocation = m_d3d12Resource->GetGPUVirtualAddress();
     m_IndexBufferView.SizeInBytes    = bufferSize;
